+++ lib/core/api/mock/splash_data.dart
