+++ lib/routes.dart
@@ -1,8 +1,8 @@
-import 'package:flubank/screens/home/home_screen.dart';
+import 'package:flubank/ui/screens/home/home_screen.dart';
+import 'package:flubank/ui/screens/login/login_screen.dart';
+import 'package:flubank/ui/screens/splash/splash_screen.dart';
 import 'package:flutter/material.dart';
 
-import 'screens/login/login_screen.dart';
-import 'screens/splash/splash_screen.dart';
 
 final Map<String, WidgetBuilder> appRoutes = {
   SplashScreen.routeName: (_) => SplashScreen(),
