+++ lib/ui/screens/home/home_screen.dart
@@ -1,6 +1,5 @@
+import 'package:flubank/ui/shared/constants.dart';
 import 'package:flutter/material.dart';
-
-import '../../constants.dart';
 
 class HomeScreen extends StatelessWidget {
   static String routeName = 'homeScreen';
