+++ lib/ui/screens/login/login_screen.dart
@@ -1,0 +1,24 @@
+import 'package:flubank/ui/screens/login/widgets/content.dart';
+import 'package:flutter/material.dart';
+
+class LoginScreen extends StatelessWidget {
+  static String routeName = 'loginPage';
+  const LoginScreen({Key key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        leading: IconButton(
+          icon: Icon(Icons.arrow_back_ios),
+          onPressed: () => Navigator.pop(context),
+        ),
+        title: Text('Inicio de sesión'),
+      ),
+      body: Padding(
+        padding: const EdgeInsets.all(8.0),
+        child: LoginScreenContent(),
+      ),
+    );
+  }
+}
