--- lib/widgets/default_button.dart
@@ -1,30 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../constants.dart';
-
-class AppDefaulButton extends StatelessWidget {
-  final Function onTap;
-  final String text;
-
-  const AppDefaulButton({Key key, @required this.onTap, @required this.text})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: double.infinity,
-      height: 50,
-      child: FlatButton(
-        onPressed: onTap,
-        color: AppColors.mainColor,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(20),
-        ),
-        child: Text(
-          text,
-          style: TextStyle(color: Colors.white, fontSize: 20),
-        ),
-      ),
-    );
-  }
-}
